pyproject.toml
[project]
name = "pm-agents"
version = "0.1.0"
description = "Multi-agent PM brainstorming system using LangGraph"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain-anthropic>=0.3.22",
    "langgraph>=0.6.11",
    "python-dotenv>=1.2.1",
    "streamlit>=1.50.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pm_agents"]

[project.scripts]
pm-agents = "pm_agents.workflow:run"
